feat: add more ships
