+++ game/Assets/Prefabs/EnemyProjectile.prefab
@@ -98,7 +98,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 931a81a8cf7fb38459322c9d1644fcb3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  speed: 7
+  speed: 3
 --- !u!61 &998900121650410138
 BoxCollider2D:
   m_ObjectHideFlags: 0
