+++ game/Assets/Scripts/PlayerController.cs
@@ -7,7 +7,6 @@ public class PlayerController : MonoBehaviour
 {
     public float moveSpeed = 5f;
     public float hInput;
-    public GameObject explosionPrefab;
     
     // Start is called before the first frame update
     void Start()
@@ -24,12 +23,4 @@ public class PlayerController : MonoBehaviour
         transform.Translate(Vector2.right * (Time.deltaTime * hInput * moveSpeed));
     }
 
-    private void OnTriggerEnter2D(Collider2D other)
-    {
-        if (other.gameObject.CompareTag("Enemy"))
-        {
-            Instantiate(explosionPrefab, transform.position, Quaternion.identity);
-            Destroy(gameObject);
-        }
-    }
 }
