+++ game/Assets/Scripts/PlayerLives.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8c6d788018a1392419f0a5c29ea6f05e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
